requirements.txt
streamlit
flask
pillow
requests
gunicorn  